quantum-crypt-encryption/requirements.txt
pqcrypto
boto3
botocore
awscli
cryptography
